+++ sock.c
@@ -84,7 +84,7 @@ status sock_addr_get_text(sock_addr_handle addr, char* text, size_t text_sz)
 				   a_buf, sizeof(a_buf)))
 		return error_errno("inet_ntop");
 
-	if (sprintf(p_buf, ":%hu", ntohs(addr->sa.sin_port)) < 0)
+	if (sprintf(p_buf, ":%d", (int) ntohs(addr->sa.sin_port)) < 0)
 		return error_errno("sock_addr_get_text");
 
 	if (strlen(a_buf) + strlen(p_buf) >= text_sz)
